# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.AddCommentRequest:
    properties:
      content:
        example: Great product!
        maxLength: 1000
        minLength: 1
        type: string
      product_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      rating:
        example: 5
        maximum: 5
        minimum: 1
        type: integer
    required:
    - content
    - product_id
    - rating
    type: object
...
  handlers.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  handlers.CreateProductRequest:
    properties:
      category:
        example: Electronics
        maxLength: 100
        minLength: 1
        type: string
      description:
        example: Latest iPhone with A17 Pro chip
        maxLength: 1000
        minLength: 1
        type: string
      image_url:
        example: https://example.com/image.jpg
        type: string
      name:
        example: iPhone 15 Pro
        maxLength: 255
        minLength: 1
        type: string
      price:
        example: 999.99
        minimum: 0.01
        type: number
      stock:
        example: 50
        minimum: 0
        type: integer
    required:
    - category
    - description
    - name
    - price
    - stock
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        example: user@example.com
        maxLength: 255
        type: string
      password:
        example: password123
        maxLength: 128
        minLength: 1
        type: string
    required:
    - email
    - password
    type: object
...
  handlers.StandardErrorResponse:
    properties:
      error:
        example: Invalid email or password
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  handlers.UpdateCommentRequest:
    properties:
      content:
        example: Updated comment
        maxLength: 1000
        minLength: 1
        type: string
      rating:
        example: 4
        maximum: 5
        minimum: 1
        type: integer
    type: object
...
  handlers.UpdateProductRequest:
    properties:
      category:
        example: Electronics
        maxLength: 100
        minLength: 1
        type: string
      description:
        example: Latest iPhone with A17 Pro chip
        maxLength: 1000
        minLength: 1
        type: string
      image_url:
        example: https://example.com/image.jpg
        type: string
      name:
        example: iPhone 15 Pro
        maxLength: 255
        minLength: 1
        type: string
      price:
        example: 999.99
        minimum: 0.01
        type: number
      stock:
        example: 50
        minimum: 0
        type: integer
    type: object
...
  models.CartItem:
    properties:
      cart:
        allOf:
        - $ref: '#/definitions/models.ShoppingCart'
        description: Relationships
      cart_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      rating:
        description: 1-5 star rating
        type: integer
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Discount:
    properties:
      category:
        description: Nullable for product-specific discounts
        type: string
      created_at:
        type: string
      discount_type:
        description: '''percentage'', ''fixed_amount'''
        type: string
      discount_value:
        description: Percentage (0-100) or fixed amount
        type: number
      end_date:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      max_discount_amount:
        description: Maximum discount amount (for percentage)
        type: number
      min_order_amount:
        description: Minimum order amount to apply discount
        type: number
      product:
        allOf:
        - $ref: '#/definitions/models.Product'
        description: Relationships
      product_id:
        description: Nullable for category-wide discounts
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      usage_count:
        type: integer
      usage_limit:
        description: 0 = unlimited
        type: integer
    type: object
  models.Favorite:
    properties:
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: string
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      status:
        type: string
      total:
        type: number
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      order:
        allOf:
        - $ref: '#/definitions/models.Order'
        description: Relationships
      order_id:
        type: string
      price:
        type: number
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  models.Product:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      category:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      description:
        type: string
      discounts:
        items:
          $ref: '#/definitions/models.Discount'
        type: array
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      order_items:
        description: Relationships
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      price:
        type: number
      product_views:
        items:
          $ref: '#/definitions/models.ProductView'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      stock:
        type: integer
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      updated_at:
        type: string
      upvotes:
        items:
          $ref: '#/definitions/models.Upvote'
        type: array
      user_interactions:
        items:
          $ref: '#/definitions/models.UserInteraction'
        type: array
    type: object
  models.ProductView:
    properties:
      created_at:
        description: Changed from Timestamp to CreatedAt for consistency
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      session_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Recommendation:
    properties:
      algorithm_type:
        description: '''collaborative'', ''content_based'', ''hybrid'''
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      score:
        type: number
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
...
  models.ShoppingCart:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      created_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.Tag:
    properties:
      color:
        description: Hex color code
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        description: Relationships
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
  models.Upvote:
    properties:
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
  models.User:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      email:
        type: string
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      id:
        type: string
      name:
        type: string
      orders:
        description: Relationships
        items:
          $ref: '#/definitions/models.Order'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      shopping_cart:
        $ref: '#/definitions/models.ShoppingCart'
      updated_at:
        type: string
      upvotes:
        items:
          $ref: '#/definitions/models.Upvote'
        type: array
      user_interactions:
        items:
          $ref: '#/definitions/models.UserInteraction'
        type: array
    type: object
  models.UserInteraction:
    properties:
      created_at:
        description: Changed from Timestamp to CreatedAt for consistency
        type: string
      id:
        type: string
      interaction_type:
        description: '''view'', ''cart_add'', ''purchase'', ''like'''
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      session_id:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@bachelor-ecommerce.com
    name: API Support
    url: http://www.swagger.io/support
  description: ML-Powered E-commerce Platform API with advanced recommendation system,
    intelligent search, and comprehensive analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bachelor E-commerce API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      summary: User login
      tags:
      - Authentication
...
  /comments:
    post:
      consumes:
      - application/json
      description: Add a comment and rating to a product
      parameters:
      - description: Comment to add
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Comment added successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add product comment
      tags:
      - Comments
  /comments/{comment_id}:
    delete:
      consumes:
      - application/json
      description: Delete a user's own comment
      parameters:
      - description: Comment ID (UUID)
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comment deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid comment ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Comment not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update a user's own comment
      parameters:
      - description: Comment ID (UUID)
        in: path
        name: comment_id
        required: true
        type: string
      - description: Updated comment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Comment updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Not authorized to update this comment
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Comment not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update comment
      tags:
      - Comments
  /comments/{product_id}:
    get:
      consumes:
      - application/json
      description: Get paginated list of comments for a product
      parameters:
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comments retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid product ID
          schema:
            additionalProperties: true
            type: object
      summary: Get product comments
      tags:
      - Comments
...
  /products:
    get:
      consumes:
      - application/json
      description: Get a paginated list of products with optional filtering and sorting
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Search in name and description
        in: query
        name: search
        type: string
      - default: '"created_at"'
        description: Sort field (price, name, created_at)
        in: query
        name: sort
        type: string
      - default: '"desc"'
        description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Products retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product in the catalog (admin access required)
      parameters:
      - description: Product creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body or validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product from the catalog (admin access required)
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid product ID or product has order history
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific product
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product retrieved successfully
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid product ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
      summary: Get product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update an existing product in the catalog (admin access required)
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Product update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body or product ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - Products
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...